This is probably my longest commit:
The django-threaded-comment app wasn't what we were looking for, so I forked it (and the default django comment system)
Create a new app because I wanted to make it independent, but I failed doing this, so ... yeah, we could merge it back
Modification were done to the comments view, model, template, form, tags, admin etc... So this become totally different from
the original app.
So now, we have the comment can be rendered

Next step will be to force authentication before commenting and a view for reply. Then I will add update and edit.